refactor(runtime): lift the field selector to a crate-level module
Move the field-selector resolver out of `alert_pipeline` into a shared
`rsigma_runtime::selector` module so post-engine sink layers can resolve
values out of an `EvaluationResult` without depending on the alert
pipeline. The `alert_pipeline` module keeps a re-export, so
`rsigma_runtime::Selector` and `rsigma_runtime::alert_pipeline::Selector`
are both unchanged. Behavior-neutral.
